move frontend schema elements (those just present on the frontend) to a separate file
